+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.0.1 (2021-??-??)
 
+* **feature:** Pagination is added to the Incidents page ([CE-1499](https://jira.rfcx.org/browse/CE-1499))
 * **feature:** User is able to close the incident ([CE-1482](https://jira.rfcx.org/browse/CE-1482))
 * **bug:** Fixed the player title label. Fixed the alignment of the navbar items ([CE-1480](https://jira.rfcx.org/browse/CE-1480))
 * **setup:** Updated the logic to return the dates difference into the dates util ([CE-1475](https://jira.rfcx.org/browse/CE-1475))